Fix merge issues, like repeated feature or stepdefs files
